Task 7: Multiple transaction is added in the form of list
